+++ Car Sharing/task/src/entity/Car.java
@@ -1,11 +1,10 @@
-package entitny;
+package entity;
 
-public class Company {
-
+public class Car {
     private int id;
     private String name;
 
-    public Company(int id, String name) {
+    public Car(int id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -24,12 +23,5 @@ public class Company {
 
     public void setName(String name) {
         this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return
-                id + ". " +
-                        name;
     }
 }
