+++ Car Sharing/task/src/entity/Company.java
@@ -1,10 +1,11 @@
-package entitny;
+package entity;
 
-public class Car {
+public class Company {
+
     private int id;
     private String name;
 
-    public Car(int id, String name) {
+    public Company(int id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -23,5 +24,12 @@ public class Car {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return
+                id + ". " +
+                        name;
     }
 }
