+++ Car Sharing/task/src/system/CarsDAOImpl.java
@@ -1,5 +1,30 @@
 package system;
 
-public class CarsDAOImpl {
+import entity.Car;
 
+import java.util.List;
+
+public class CarsDAOImpl implements ICarsDAO {
+    public CarsDAOImpl() {
+        creatIfNoExist();
+    }
+
+     void creatIfNoExist(){
+
+    }
+
+    @Override
+    public void create(Car car) {
+
+    }
+
+    @Override
+    public Car get(int id) {
+        return null;
+    }
+
+    @Override
+    public List<Car> getAll() {
+        return null;
+    }
 }
