+++ Car Sharing/task/src/system/ICompanyDAO.java
@@ -1,6 +1,6 @@
 package system;
 
-import entitny.Company;
+import entity.Company;
 
 import java.util.List;
